Introduce preliminary support for GPT-4 Turbo
A new model Gpt4Turbo has been introduced in the ChatCompletionModels class with a tokens limit set to 4096. Corresponding updates have been made to the integration test cases replacing the previous model with this new one. Furthermore, the project version has been updated to 2.9.2 in the Directory.Build.props file