Update code and project version.
In `ServiceCollectionExtensions.cs`, the order of arguments in `AddChatGptEntityFrameworkIntegration` function is changed. This was done to ensure `optionsAction` is not null before it's used. Also, validation for `optionsAction` is moved right after validating `configuration`.

In `Directory.Build.props`, the version number has been updated from `4.0.0-alpha` to `4.0.1-alpha`. This indicates a minor revision or update in the project.